Fix error handling
